refactor: move kakaologic from MemberController to AuthController
